// src/lib/types/userDataTypes.ts
export interface UserProfileData {
  id: number;
  name: string;
  email: string;
  phone: string;
  image: string;
  last_login: null;
  first_name: string;
  last_name: string;
  date_joined: Date;
  created: Date;
  updated: Date;
  uip: null;
  bio: string;
  cover: string;
  contract_email: null;
  contract_number: null;
  groups: any[];
  user_permissions: any[];
}